+++ packages/examples/src/layout/footer.tsx
@@ -4,7 +4,7 @@ export const Footer: FC = () => {
   return (
     <footer className="footer fixed-bottom ">
       <div className="container-fluid">
-        <p className="text-center">&copy; @react-simga - 2022</p>
+        <p className="text-center">&copy; react-sigma - 2022</p>
       </div>
     </footer>
   );